add Message component and style it.
