+++ src/master.css
@@ -7,8 +7,36 @@ body {
   background: #3167ad;
   color: #ffffff;
   text-align: center;
-  padding: 10px;
-  font-size: 17px;
+  padding: 15px;
+  font-size: 20px;
   box-shadow: 0 0 5px black;
   font-weight: bold;
+}
+
+.message-container {
+  margin: 20px;
+  padding: 20px;
+  background: RGBA(145, 206, 255, .8);
+  border-radius: 20px;
+  box-shadow: 0 0 5px black;
+  display: flex;
+  justify-content: space-around;
+}
+
+.message-container h3 {
+  font-size: 15px;
+}
+
+.btn-status {
+  width: 10%;
+  border-radius: 10px;
+  cursor: pointer;
+  color: white;
+  background: RGBA(49, 103, 173, .8);
+  font-size: 15px;
+}
+
+.btn-status:hover {
+  color: RGB(49, 103, 173);
+  background: white;
 }
